implement the new view controller
